Rename projectRoutes.ts
